fix: address P2 finding #023 — extract raw DB queries from routes to service modules
- Moved prisma.request311.findMany to fetchBlockRequests() in block service
- Created server/services/metrics.ts with fetchCommunityMetrics() and
  buildCommunityMetricsResponse() extracted from the metrics route
- Routes now delegate to services for all data access and business logic
